requirements.txt
fastapi
uvicorn
beautifulsoup4
playwright
python-dotenv
